Use DelegationNetwork; delete DelegationInfo; pull ToJSON instance from Arbitrary.
